+++ Source/Runtime/ProgramCore/Public/Types/xxHash/xxHashInclude.hpp
@@ -35,22 +35,22 @@ namespace xxHash
 template <typename CharType>
 CONST_EXPR uint32 hashString(const CharType *str, uint32 seed)
 {
-    return xxh32::hash(str, TCharStr::length(str) * sizeof(CharType), seed);
+    return xxh32::hash(str, uint32(TCharStr::length(str) * sizeof(CharType)), seed);
 }
 template <typename CharType>
 CONST_EXPR uint64 hashString(const CharType *str, uint64 seed)
 {
-    return xxh64::hash(str, TCharStr::length(str) * sizeof(CharType), seed);
+    return xxh64::hash(str, uint32(TCharStr::length(str) * sizeof(CharType)), seed);
 }
 template <typename CharType>
 CONST_EXPR uint32 hashString(const CharType *str, uint32 len, uint32 seed)
 {
-    return xxh32::hash(str, len * sizeof(CharType), seed);
+    return xxh32::hash(str, uint32(len * sizeof(CharType)), seed);
 }
 template <typename CharType>
 CONST_EXPR uint64 hashString(const CharType *str, uint64 len, uint64 seed)
 {
-    return xxh64::hash(str, len * sizeof(CharType), seed);
+    return xxh64::hash(str, uint32(len * sizeof(CharType)), seed);
 }
 
 FORCE_INLINE uint32 hashString(const String &str, uint32 seed)
@@ -61,7 +61,7 @@ FORCE_INLINE uint32 hashString(const String &str, uint32 seed)
     }
     else
     {
-        return xxh32::hash(str.getChar(), str.length() * sizeof(String::value_type), seed);
+        return xxh32::hash(str.getChar(), uint32(str.length() * sizeof(String::value_type)), seed);
     }
 }
 FORCE_INLINE uint64 hashString(const String &str, uint64 seed)
@@ -72,7 +72,7 @@ FORCE_INLINE uint64 hashString(const String &str, uint64 seed)
     }
     else
     {
-        return xxh64::hash(str.getChar(), str.length() * sizeof(String::value_type), seed);
+        return xxh64::hash(str.getChar(), uint32(str.length() * sizeof(String::value_type)), seed);
     }
 }
 } // namespace xxHash
